// sdk/src/models/channel-list-entry.ts
/*
 * Code generated by Speakeasy (https://speakeasy.com). DO NOT EDIT.
 */

import * as z from "zod/v3";
import { safeParse } from "../lib/schemas.js";
import { Result as SafeParseResult } from "../types/fp.js";
import * as types from "../types/primitives.js";
import {
  Channel,
  Channel$inboundSchema,
  Channel$Outbound,
  Channel$outboundSchema,
} from "./channel.js";
import { SDKValidationError } from "./errors/sdk-validation-error.js";
import {
  Message,
  Message$inboundSchema,
  Message$Outbound,
  Message$outboundSchema,
} from "./message.js";

/**
 * `listChannels` returns each channel as a tuple of the channel itself
 *
 * @remarks
 * plus its pinned messages and member-ID list.
 */
export type ChannelListEntry = {
  channel: Channel;
  pinnedMessages?: Array<Message> | undefined;
  /**
   * IDs of members of the channel.
   */
  users?: Array<string> | undefined;
};

/** @internal */
export const ChannelListEntry$inboundSchema: z.ZodType<
  ChannelListEntry,
  z.ZodTypeDef,
  unknown
> = z.object({
  channel: Channel$inboundSchema,
  pinnedMessages: types.optional(z.array(Message$inboundSchema)),
  users: types.optional(z.array(types.string())),
});
/** @internal */
export type ChannelListEntry$Outbound = {
  channel: Channel$Outbound;
  pinnedMessages?: Array<Message$Outbound> | undefined;
  users?: Array<string> | undefined;
};

/** @internal */
export const ChannelListEntry$outboundSchema: z.ZodType<
  ChannelListEntry$Outbound,
  z.ZodTypeDef,
  ChannelListEntry
> = z.object({
  channel: Channel$outboundSchema,
  pinnedMessages: z.array(Message$outboundSchema).optional(),
  users: z.array(z.string()).optional(),
});

export function channelListEntryToJSON(
  channelListEntry: ChannelListEntry,
): string {
  return JSON.stringify(
    ChannelListEntry$outboundSchema.parse(channelListEntry),
  );
}
export function channelListEntryFromJSON(
  jsonString: string,
): SafeParseResult<ChannelListEntry, SDKValidationError> {
  return safeParse(
    jsonString,
    (x) => ChannelListEntry$inboundSchema.parse(JSON.parse(x)),
    `Failed to parse 'ChannelListEntry' from JSON`,
  );
}